server_flask/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
pymongo>=4.8,<5.0
python-dotenv>=1.0,<2.0
